requirements.txt
google-api-python-client==2.108.0
python-dotenv==1.0.0
rich==13.7.0
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
